Corrected the type of variable of the latitude and longitude attributes of the shores table

+++ database/migrations/2020_07_04_122119_create_shores_table.php
@@ -19,8 +19,8 @@ class CreateShoresTable extends Migration
             $table->string('code', 2);
             $table->foreignId('country_id')->constrained();
             $table->foreignId('fwunit_id')->nullable()->constrained('field_work_units');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->double('latitude', 9, 6);
+            $table->double('longitude', 9, 6);
             $table->timestamps();
         });
     }
